requirements.txt
python-dateutil
aiocron
discord.py
sqlalchemy
